Add new test_pixel_sky_conversions module
