+++ photutils/aperture/tests/test_photometry.py
@@ -21,9 +21,8 @@ from photutils.aperture.rectangle import (RectangularAnnulus,
                                           RectangularAperture,
                                           SkyRectangularAnnulus,
                                           SkyRectangularAperture)
-from photutils.datasets import make_4gaussians_image, make_gwcs, make_wcs
-from photutils.utils._optional_deps import (HAS_GWCS, HAS_MATPLOTLIB,
-                                            HAS_REGIONS)
+from photutils.datasets import make_4gaussians_image, make_wcs
+from photutils.utils._optional_deps import HAS_MATPLOTLIB, HAS_REGIONS
 
 APERTURE_CL = [CircularAperture,
                CircularAnnulus,
@@ -667,65 +666,6 @@ def test_elliptical_bbox():
 
     ap = EllipticalAperture((50, 50), a=a, b=b, theta=np.deg2rad(90.0))
     assert ap.bbox.shape == (2 * a, 2 * b)
-
-
-@pytest.mark.skipif(not HAS_GWCS, reason='gwcs is required')
-@pytest.mark.parametrize('wcs_type', ['wcs', 'gwcs'])
-def test_to_sky_pixel(wcs_type):
-    data = make_4gaussians_image()
-
-    if wcs_type == 'wcs':
-        wcs = make_wcs(data.shape)
-    elif wcs_type == 'gwcs':
-        wcs = make_gwcs(data.shape)
-
-    ap = CircularAperture(((12.3, 15.7), (48.19, 98.14)), r=3.14)
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.r, ap2.r)
-
-    ap = CircularAnnulus(((12.3, 15.7), (48.19, 98.14)), r_in=3.14,
-                         r_out=5.32)
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.r_in, ap2.r_in)
-    assert_allclose(ap.r_out, ap2.r_out)
-
-    ap = EllipticalAperture(((12.3, 15.7), (48.19, 98.14)), a=3.14, b=5.32,
-                            theta=np.deg2rad(103.0))
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.a, ap2.a)
-    assert_allclose(ap.b, ap2.b)
-    assert_allclose(ap.theta, ap2.theta)
-
-    ap = EllipticalAnnulus(((12.3, 15.7), (48.19, 98.14)), a_in=3.14,
-                           a_out=15.32, b_out=4.89,
-                           theta=np.deg2rad(103.0))
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.a_in, ap2.a_in)
-    assert_allclose(ap.a_out, ap2.a_out)
-    assert_allclose(ap.b_out, ap2.b_out)
-    assert_allclose(ap.theta, ap2.theta)
-
-    ap = RectangularAperture(((12.3, 15.7), (48.19, 98.14)), w=3.14, h=5.32,
-                             theta=np.deg2rad(103.0))
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.w, ap2.w)
-    assert_allclose(ap.h, ap2.h)
-    assert_allclose(ap.theta, ap2.theta)
-
-    ap = RectangularAnnulus(((12.3, 15.7), (48.19, 98.14)), w_in=3.14,
-                            w_out=15.32, h_out=4.89,
-                            theta=np.deg2rad(103.0))
-    ap2 = ap.to_sky(wcs).to_pixel(wcs)
-    assert_allclose(ap.positions, ap2.positions)
-    assert_allclose(ap.w_in, ap2.w_in)
-    assert_allclose(ap.w_out, ap2.w_out)
-    assert_allclose(ap.h_out, ap2.h_out)
-    assert_allclose(ap.theta, ap2.theta)
 
 
 def test_scalar_aperture():
